Clean up and light directory structure refactor
