+++ package.json
@@ -21,8 +21,7 @@
     "build": "yarn run clean-lib && tsc",
     "format": "prettier '{src,test}/**/*' ./*.json --write --loglevel warn",
     "lint": "eslint src/**/*.ts test/**/*.ts --fix",
-    "test": "yarn run clean-tmp && yarn run build && ts-mocha --timeout 10000 --recursive --require @babel/register --exit \"test/**/*.test.ts\"",
-    "test:temp": "ts-mocha --timeout 10000 --recursive --require @babel/register --exit \"src/**/*.test.ts\"",
+    "test": "yarn run clean-tmp && yarn run build && ts-mocha --timeout 10000 --recursive --require @babel/register --exit \"{test,src}/**/*.test.ts\"",
     "cover": "cross-env BABEL_ENV=coverage nyc --reporter=lcov --reporter=text npm test",
     "posttest": "yarn run format && yarn run lint",
     "docs:clean": "del-cli _book",
