+++ src/cli/argsParser/optionsFromParsedArgs/webpack/requireWebpackConfig.ts
@@ -8,7 +8,11 @@ import {
   isUndefined as _isUndefined
 } from 'lodash'
 import { Configuration } from 'webpack'
-import { ModuleDescriptor, WebpackConfig, WebpackConfigMode } from './types'
+import {
+  ModuleDescriptor,
+  WebpackConfig,
+  WebpackConfigMode
+} from '../../../types'
 
 /**
  * Ensures that .js extension comes first, or whichever extension is shorter
