+++ src/runner/runnerUtils/createWebpackConfig/createWebpackConfig.test.ts
@@ -6,7 +6,7 @@ import { ProgressPlugin, LoaderOptionsPlugin } from 'webpack'
 import createWebpackConfig from '.'
 import { CreateWebpackConfigOptions, MochapackWebpackConfigs } from './types'
 import * as mochapackPlugins from '../../../webpack/plugin/buildProgressPlugin'
-import { MOCHAPACK_NAME } from '../constants'
+import { MOCHAPACK_NAME } from '../../../util/constants'
 
 describe('createWebpackConfig', () => {
   const fixturesDir = resolve(
